engine: extract dispatch::reaction_windows (#159)
